+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Features
 
+* **css:** Add Picto Card component (#154, #155)
 * **css:** Add Call Out component (#153)
 * **css:** Update button styles (#167)
 * **css:** Add custom .mzp-t-firefox global theme class (#169)