+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# HEAD
+# 2.0.0
 
 ## Features